Refactor (APIMF-3491): extracted object map property parsing
