+++ amf-aml/shared/src/main/scala/amf/aml/internal/parse/instances/DialectInstanceParser.scala
@@ -46,6 +46,7 @@ import amf.aml.internal.parse.instances.parser.{
   LiteralValueParser,
   LiteralValueSetter,
   ObjectCollectionPropertyParser,
+  ObjectMapPropertyParser,
   ObjectUnionParser
 }
 import amf.aml.internal.validate.DialectValidations.{
@@ -371,32 +372,6 @@ class DialectInstanceParser(val root: Root)(implicit override val ctx: DialectIn
     DialectExtensionParser.parse(id, propertyEntry, property, node, root, parseNestedNode)
   }
 
-  protected def findHashProperties(propertyMapping: PropertyMapping, propertyEntry: YMapEntry): Option[(String, Any)] = {
-    propertyMapping.mapTermKeyProperty().option() match {
-      case Some(propId) => Some((propId, propertyEntry.key.as[YScalar].text))
-      case None         => None
-    }
-  }
-
-  protected def checkHashProperties(node: DialectDomainElement,
-                                    propertyMapping: PropertyMapping,
-                                    propertyEntry: YMapEntry): DialectDomainElement = {
-    // TODO: check if the node already has a value and that it matches (maybe coming from a declaration)
-    propertyMapping.mapTermKeyProperty().option() match {
-      case Some(propId) =>
-        try {
-          node.set(Field(Str, ValueType(propId)),
-                   AmfScalar(propertyEntry.key.as[YScalar].text),
-                   Annotations(propertyEntry.key))
-        } catch {
-          case e: UnknownMapKeyProperty =>
-            ctx.eh.violation(DialectError, e.id, s"Cannot find mapping for key map property ${e.id}")
-            node
-        }
-      case None => node
-    }
-  }
-
   protected def parseObjectUnion[T <: DomainElement](
       defaultId: String,
       path: Seq[String],
@@ -436,37 +411,13 @@ class DialectInstanceParser(val root: Root)(implicit override val ctx: DialectIn
                                        property: PropertyMapping,
                                        node: DialectDomainElement,
                                        additionalProperties: Map[String, Any] = Map()): Unit = {
-    val nested = propertyEntry.value.as[YMap].entries.map { keyEntry =>
-      val path           = List(propertyEntry.key.as[YScalar].text, keyEntry.key.as[YScalar].text)
-      val nestedObjectId = pathSegment(id, path)
-      // we add the potential mapKey additional property
-      val keyAdditionalProperties: Map[String, Any] = findHashProperties(property, keyEntry) match {
-        case Some((k, v)) => additionalProperties + (k -> v)
-        case _            => additionalProperties
-      }
-      val parsedNode = property.nodesInRange match {
-        case range: Seq[String] if range.size > 1 =>
-          // we also add the potential mapValue property
-          val keyValueAdditionalProperties = property.mapTermValueProperty().option() match {
-            case Some(mapValueProperty) => keyAdditionalProperties + (mapValueProperty -> "")
-            case _                      => keyAdditionalProperties
-          }
-          // now we can parse the union with all the required properties to find the right node in the union
-          Some(parseObjectUnion(nestedObjectId, path, keyEntry.value, property, keyValueAdditionalProperties))
-        case range: Seq[String] if range.size == 1 =>
-          ctx.dialect.declares.find(_.id == range.head) match {
-            case Some(nodeMapping: NodeMappable) if keyEntry.value.tagType != YType.Null =>
-              Some(parseNestedNode(id, nestedObjectId, keyEntry.value, nodeMapping, keyAdditionalProperties))
-            case _ => None
-          }
-        case _ => None
-      }
-      parsedNode match {
-        case Some(dialectDomainElement) => Some(checkHashProperties(dialectDomainElement, property, keyEntry))
-        case None                       => None
-      }
-    }
-    node.withObjectCollectionProperty(property, nested.flatten, Right(propertyEntry))
+    ObjectMapPropertyParser.parse(id,
+                                  propertyEntry,
+                                  property,
+                                  node,
+                                  additionalProperties,
+                                  parseObjectUnion,
+                                  parseNestedNode)
   }
 
   protected def parseObjectPairProperty(id: String,
